% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_any_match}
\alias{row_any_match}
\title{Row Any Match}
\usage{
row_any_match(..., value, ignoreNA = TRUE, ignoreAllNA = FALSE)
}
\arguments{
\item{...}{Variables over which to check for instances equal to value (no limit on number of variables)}

\item{value}{Specified value to count; can be numeric or nominal or character}

\item{ignoreNA}{Indicates whether missing values should be
ignored while checking the number of specified values; default is ignoreNA = TRUE. However,
if all variables are missing, then a value is not returned, but is NA.
If ignoreNA = TRUE, then columns with a missing value are ignored in each row as long
as at least one variable is non-missing.
If ignoreNA = FALSE, then a value is not returned regardless of the number of missing variables,
but is NA.}

\item{ignoreAllNA}{When TRUE, indicates that a value should be returned even if all
variables are missing (a value of 0 will be returned in such cases). Default is FALSE. If ignoreAllNA is TRUE, then ignoreNA is assumed
to be and must also be TRUE.}
}
\value{
Vector of length of each input variable
}
\description{
Indicates if at least one of variables (columns) is equal to a specified value across variables within each row. Returns a value of 1 if at least one variable equals the value; 0 or NA otherwise.
}
