% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_less_count}
\alias{row_less_count}
\title{Row Less Count}
\usage{
row_less_count(..., value, maxmiss_n = 0)
}
\arguments{
\item{...}{Variable list over which to count instances equal to value (no limit on number of
variables)}

\item{value}{Specified value; must be numeric. Values greater than than the specified value
will be counted.}

\item{maxmiss_n}{Maximum number of missing values allowed in the variables list in
order to return a count; default is 0. For each row, if # of missing values <= maxmiss_n,
then the count is calculated using the remaining variables with non-missing values; if # of
missing values > maxmiss_n, then NA is returned.}
}
\value{
Vector of length of each input variable
}
\description{
For each row, returns the count of variables in the variable list (...) with values less
than a specified value. Includes options for handling missing data.
}
