% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_mean}
\alias{row_mean}
\title{Row Mean}
\usage{
row_mean(..., maxmiss_n = 0)
}
\arguments{
\item{...}{Variables to be averaged (no limit on number of variables, but must be numeric).}

\item{maxmiss_n}{Maximum number of missing values allowed in the variables list in
order to return an average; default is 0. For each row, if # of missing values <= maxmiss_n,
then the average is calculated using the remaining variables with non-missing values; if # of
missing values > maxmiss_n, then NA is returned. maxmiss_n must be less than the number of
variables.}
}
\value{
Vector of length of each input variable.
}
\description{
Calculates the average value across variables within each row. Includes an option for
handling missing data.
}
