% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_means}
\alias{row_means}
\title{Row Mean}
\usage{
row_means(..., n = 0)
}
\arguments{
\item{...}{Variables to be averaged (no limit on number of variables, but must be numeric).}

\item{n}{Maximum number of non-missing values allowed in order to compute an average; default is 0. When n > 0, the mean for each case is calculated using the remaining variables with non-missing values.}
}
\value{
Vector of length of each input variable.
}
\description{
Calculates the average across variables (columns) within each row. Includes an option for handling missing data.
}
