% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_na}
\alias{row_na}
\title{Row Missing}
\usage{
row_na(...)
}
\arguments{
\item{...}{Variables over which to count missing values (no limit on number of variables)}
}
\value{
Vector of length of each input variable
}
\description{
Counts the number of missing values across variables (columns) within each row.
}
