% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_nvalid}
\alias{row_nvalid}
\title{Row Non-Missing}
\usage{
row_nvalid(...)
}
\arguments{
\item{...}{Variables over which to count non-missing values (no limit on number of variables)}
}
\value{
Vector of length of each input variable
}
\description{
Counts the number of non-missing values across variables (columns) within each row.
}
