% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_sum}
\alias{row_sum}
\title{Row Sum}
\usage{
row_sum(..., maxmiss_n = 0)
}
\arguments{
\item{...}{Variables to be summed (no limit on number of variables, but must be numeric).}

\item{maxmiss_n}{Maximum number of missing values allowed in the variables list in
order to return a sum; default is 0. For each row, if # of missing values <= maxmiss_n,
then the sum is calculated using the remaining variables with non-missing values; if # of
missing values > maxmiss_n, then NA is returned. maxmiss_n must be less than the number of
variables.}
}
\value{
Vector of length of each input variable.
}
\description{
Calculates sum across variables within each row. Includes an option for handling missing data.
}
