% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_any_count}
\alias{row_any_count}
\title{Row Any Count}
\usage{
row_any_count(..., value, ignoreNA = TRUE, ignoreAllNA = FALSE)
}
\arguments{
\item{...}{Variables over which to count instances equal to value (no limit on number of variables)}

\item{value}{Specified value to count; can be numeric or nominal or character}

\item{ignoreNA}{Indicates whether missing values should be
ignored while counting the number of specified values; default is ignoreNA = TRUE. However,
if all variables are missing, then a count is not computed, but is NA.
If ignoreNA = TRUE, then columns with a missing value are ignored in each row as long
as at least one variable is non-missing.
If ignoreNA = FALSE, then a count is not computed regardless of the number of missing variables,
but is NA.}

\item{ignoreAllNA}{When TRUE, indicates that a count should be computed even if all
variables are missing. Default is FALSE. If ignoreAllNA is TRUE, then ignoreNA is assumed
to be and must also be TRUE.}
}
\value{
Vector of length of each input variable
}
\description{
Returns the count of variables in the variable list (...) equal to a specified value. Includes options for handling missing data.
}
