% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_any_match}
\alias{row_any_match}
\title{Row Any Match}
\usage{
row_any_match(..., values, maxmiss_n = 0)
}
\arguments{
\item{...}{List of variables over which to check for a match (no limit
on number of variables)}

\item{values}{Specified value or values to count; can be numeric, nominal, or character. Must
be placed in quotes.}

\item{maxmiss_n}{Maximum number of missing values allowed in the variables list in
order to return a value; default is 0. For each row, if # of missing values <= maxmiss_n,
then whether (1) or not (0) there is a match is returned using the remaining variables with
non-missing values; if # of missing values > maxmiss_n, then NA is returned.}
}
\value{
Vector of length of each input variable

Vector of length of each input variable
}
\description{
Checks whether the value of any variable in a variable list matches a specified value or values.
Returns a value of 1 if at least one variable in the variable list (...) is equal to a
specified value or values; 0 or NA otherwise. Includes options for handling missing data.
}
