% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgen.R
\name{row_sum}
\alias{row_sum}
\title{Row Sum}
\usage{
row_sum(..., n = 0)
}
\arguments{
\item{...}{Variables to be summed (no limit on number of variables, but must be numeric).}

\item{n}{Maximum number of non-missing values allowed in order to compute a sum; default is 0. Missing values are treated as 0.}
}
\value{
Vector of length of each input variable.
}
\description{
Sums values across variables (columns) within each row. Includes an option for handling missing data.
}
